Littlelemon/requirements.txt
# Core
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
django-filter>=25.1,<26.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0
djoser>=2.2,<3.0

# Security & CORS
django-cors-headers>=4.4,<5.0
python-decouple>=3.8,<4.0

# Caching & Redis
django-redis>=5.4,<6.0
redis>=5.0,<6.0

# Background Tasks
celery[redis]>=5.4,<6.0
django-celery-beat>=2.6,<3.0

# WebSockets
channels>=4.2,<5.0
channels-redis>=4.3,<5.0

# Logging & Monitoring
structlog>=24.0

# Database
psycopg2-binary>=2.9,<3.0

# Utilities
python-dotenv>=1.0,<2.0